First hack at merging in pbs_jobnames
The problem is... testing helper functions in a script isn't easy if there
isn't a python module associated with it to import. In this commit I've created
a python module that symlinks to the bin/qbatch file because I wanted to keep
all the code in one file.

I've also used the mock package to mock out the call to qstat -x when testing.
It's not particularly elegant in that it's not very blackbox. The other option
is to extract a function that calls qstat and mock that during testing.

Also, I'm not particularly happy with the use of exceptions here, but I also
don't think it matters greatly.

This addresses #46. Thoughts?